fix(Prefabs): allow teleport to same location as previous location
Zinnia 1.16.0 has some changes to the SurfaceLocator and the
TransformPropertyApplier that allows both components to still operate
if the previous location is the same as the current loaction.

This limitation in Zinnia 1.15.0 meant that you could not teleport to
the same position that you were currently on, which was fine in most
cases but there are some occasions where you do want to teleport
back to where you're currently standing. This is now the default
setting, but its still possible to switch the settings to have it
work the way it did before.