feat(prefab): add ability to customise snap to floor thresholds
The snap to floor threshold has now been exposed in the facade and
the snap to floor mechanism has been updated so two Surface Change
Actions are now used. The first one determines if a snap to the
nearest floor can even take place (if the distance between the old
and new floor exceeds the threshold). The second one determines
if the snap teleport will blink the view.

This enables now certain small changes in floor height not to
teleport the target at all and to only blink the view when the floor
height has changed substantially enough.