fix(TeleporterConfigurator): configure offsets when enabled
The Surface Locator offsets were not being configured and setup when
the configurator was enabled and the code that did the configuration
required the entire facade to be disabled then re-enabled which would
cause an infinite loop if called in the OnEnable.

This has been fixed by not needing the whole Facade to be disabled and
instead a simple flag has been added that delays the setting of the
threshold until the offsets have been set up, which means the offset
set up can be run in the OnEnable.