fix(SnapToFloor): disable camera color overlay by default
If the CameraColorOverlay object is not disabled by default then
the camera will do an initial fade if the destination offset is higher
than 0 for the Y axis. The process is already there to enable the
CameraColorOverlay after the first snap to floor logic run but the
GameObject was left enabled by default so it will do the fade on the
first snap to floor logic run.