fix(SnapToFloor): re-enable the camera color overlay component
The CameraColorOverlay component had been disabled to prevent an
initial blink when the teleporter started at an offset height but
this caused the blink to be always disabled if no offset height was
provided.

The component is now enabled by default and at a later date a further
fix for the offset height blink will be added.