fix(prefabs): ensure snap to floor enabled turns off correct object
The `Snap To Floor Enabled` property was not working correctly as it
was turning off the entire Snap To Floor prefab meaning the blink
logic was also not working as the blink/fade logic is contained within
the Snap To Floor prefab.

The fix is to just turn off the actual surface location logic so the
blink is still accessible.