fix(Prefabs): ensure snap to floor moves along sloped objects
The Snap To Floor logic was not moving along sloped objects because
the surface change action was only triggering once when the surface
changed, but if the position changed over the same object then
it was not triggering a new change event.

This has been fixed now by an update to the Zinnia Action class
that allows the action value to be reset so it is now reset after
the activated action is emitted in the surface change action so
the next time it is checked, the action can still be emitted.

The MomentProcessor within the teleporter prefabs is also exposed
onto a top level MomentProcess on the teleporter prefabs so it
can be easily added to another MomentProcessor.