feat(prefabs): add pause/resume snap to floor logic
The SnapToFloor prefab now has two EmptyEventProxyEmitters that can
just pause or resume the Snap To Floor logic without turning it off
completely, which causes a complete reset.

The pause logic, just turns off the component that physically does
the moving of the target and turns off the camera fade logic.

The internals of the prefab have also been tidied up into an
`Internal` GameObject to keep it neater.